+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## [0.1.2](https://github.com/side-xp/node-facile/compare/v0.1.1...v0.1.2) (2025-12-16)
+
+
+### Features
+
+* Added get/set field value for inputs ([93d00c6](https://github.com/side-xp/node-facile/commit/93d00c6ef8c07cca1ffedbc1a6760b5f717b473c))
+* Added math utilities ([ccde37b](https://github.com/side-xp/node-facile/commit/ccde37b64173db4e90b5ddc5c5c2cdf3ec5ee86d))
+* onChange event utility ([0a1ee5e](https://github.com/side-xp/node-facile/commit/0a1ee5ef202cc93b26d4d8c74c894d98cd144a6f))
+
 ## [0.1.1](https://github.com/side-xp/node-facile/compare/v0.1.0...v0.1.1) (2025-12-15)
 
 ## [0.1.0](https://github.com/side-xp/node-facile/compare/v0.0.4...v0.1.0) (2025-12-15)